+++ misc/longest_common_subseq.h
@@ -18,8 +18,8 @@ vector<T> longest_common_subseq(
     const vector<T>& a,
     const vector<T>& b) {
   vector length(size(a) + 1, vector<int>(size(b) + 1));
-  for (int i = 0; i <= size(a); i++) {
-    for (int j = 0; j <= size(b); j++) {
+  for (int i = 0; i <= (int)size(a); i++) {
+    for (int j = 0; j <= (int)size(b); j++) {
       if (i > 0) length[i][j] = max(length[i][j], length[i-1][j]);
       if (j > 0) length[i][j] = max(length[i][j], length[i][j-1]);
       if (i > 0 && j > 0 && a[i-1] == b[j-1]) {
@@ -28,7 +28,7 @@ vector<T> longest_common_subseq(
     }
   }
   vector<T> lcs;
-  for (int i = size(a), j = size(b); i > 0 && j > 0; ) {
+  for (int i = (int)size(a), j = (int)size(b); i > 0 && j > 0; ) {
     if (a[i-1] == b[j-1]) {
       lcs.push_back(a[i-1]);
       i--; j--;
@@ -41,3 +41,4 @@ vector<T> longest_common_subseq(
   reverse(begin(lcs), end(lcs));
   return lcs;
 }
+
